Update plist.m3u
Updated AKTV sources.

If possible, please follow the original source at aktv.space/live.m3u for uninterrupted viewing. If said playlist breaks, a quick refresh on the channels will do.

Reminder: This source is for your PERSONAL use and it's provided for free. If you paid to access this, REQUEST A REFUND.